+++ src/labkey/Assay.ts
@@ -14,9 +14,40 @@
  * limitations under the License.
  */
 import { buildURL } from './ActionURL'
-import { AjaxHandler, request } from './Ajax'
+import { request } from './Ajax'
 import { appendFilterParams, IFilter } from './filter/Filter'
-import { applyTranslated, displayAjaxErrorResponse, getCallbackWrapper, getOnFailure, getOnSuccess, merge } from './Utils'
+import { QueryColumn } from './query/types';
+import {
+    applyTranslated,
+    displayAjaxErrorResponse,
+    getCallbackWrapper,
+    getOnFailure,
+    getOnSuccess,
+    merge,
+    RequestCallbackOptions
+} from './Utils'
+
+/**
+ * Applies arguments for backwards compatible function signature support.
+ * @hidden
+ * @private
+ */
+function applyArguments(args: IArguments, options: GetAssaysOptions, parameter?: string): GetAssaysOptions {
+    let _options: GetAssaysOptions;
+
+    if (args.length > 1) {
+        _options = {
+            success: args[0],
+            failure: args[1],
+            parameters: parameter ? { [parameter]: args[2] } : {},
+            containerPath: parameter ? args[3] : args[2],
+        };
+    } else {
+        _options = options;
+    }
+
+    return _options;
+}
 
 /**
  * Gets all assays
@@ -51,32 +82,17 @@ import { applyTranslated, displayAjaxErrorResponse, getCallbackWrapper, getOnFai
  * @param options
  * @see [[AssayDesign]]
  */
-export function getAll(options: IGetAssaysOptions) {
-
-    if (arguments.length > 1) {
-        options = {
-            success: arguments[0],
-            failure: arguments[1],
-            parameters: {},
-            containerPath: arguments[2]
-        };
-    }
-
-    getAssays(options);
+export function getAll(options: GetAssaysOptions): XMLHttpRequest {
+    return getAssays(applyArguments(arguments, options));
 }
 
-export interface IGetAssaysOptions {
-    /**The container path in which the requested Assays are defined.
-     * If not supplied, the current container path will be used.*/
+export interface GetAssaysOptions extends RequestCallbackOptions<AssayDesign[]> {
+    /**
+     * The container path in which the requested Assays are defined.
+     * If not supplied, the current container path will be used.
+     */
     containerPath?: string
-    /**Function called when execution of the "getAll" function fails.*/
-    failure?: Function
     parameters?: any
-    /**The scope to be used for the success and failure callbacks*/
-    scope?: any
-    /**Function called when the "getAll" function executes successfully.
-     * Will be called with the argument: [[AssayDesign]][].*/
-    success: ( assays: AssayDesign[] ) => void
 }
 
 enum AssayLink {
@@ -90,7 +106,6 @@ enum AssayLink {
     RUNS = 'runs'
 }
 
-
 export interface AssayDesign {
     /** The name of the assay. */
     name: string
@@ -102,12 +117,12 @@ export interface AssayDesign {
      * Map containing name/value pairs.  Typically contains three entries for three domains (batch, run and results).
      * Each domain is associated with an array of objects that each describe a domain field.
      */
-    domains: Map<string, any /*FieldMetaData*/[]>;
+    domains: { [domainName:string]: QueryColumn };
     /**
      * An mapped enumeration of domain types to domain names. Useful when attempting to find a domain by type.
      * The value is a domain name which can be used as a key lookup into the "domain" object.
      */
-    domainTypes: Map<string, string>;
+    domainTypes: { [domainType:string]: string };
     /** The unique ID of the assay. */
     id: number;
     /** The name of the action used for data import. */
@@ -122,7 +137,6 @@ export interface AssayDesign {
     projectLevel: boolean;
     /** Query schema name of this assay protocol. e.g. 'assay.General.MyAssay' */
     protocolSchemaName: string;
-    // sampleRuns: List<number>;
     /** URL for generating an Excel template for importing data into this assay. */
     templateLink: string;
     /** The name of the assay type.  Example:  "ELISpot". */
@@ -133,33 +147,28 @@ export interface AssayDesign {
  * @hidden
  * @private
  */
-function getAssays(options: IGetAssaysOptions): void {
-
-    if (arguments.length > 1) {
-        options = {
-            success: arguments[0],
-            failure: arguments[1],
-            parameters: arguments[2],
-            containerPath: arguments[3]
-        };
-    }
-
+function getAssays(options: GetAssaysOptions): XMLHttpRequest {
     moveParameter(options, 'id');
     moveParameter(options, 'type');
     moveParameter(options, 'name');
 
-    request({
+    return request({
         url: buildURL('assay', 'assayList.api', options.containerPath),
         method: 'POST',
         jsonData: options.parameters,
-        success: getSuccessCallbackWrapper(getOnSuccess(options), options.scope),
+        success: getCallbackWrapper(
+            getOnSuccess(options),
+            options.scope,
+            false,
+            (data) => data.definitions
+        ),
         failure: getCallbackWrapper(getOnFailure(options), options.scope, true),
         scope: options.scope || this
     });
 }
 
-export interface IGetByIdOptions extends IGetAssaysOptions {
-    /**Unique integer ID for the assay.*/
+export interface GetByIdOptions extends GetAssaysOptions {
+    /** Unique integer ID for the assay. */
     id: number
 }
 
@@ -168,25 +177,12 @@ export interface IGetByIdOptions extends IGetAssaysOptions {
  * @param options
  * @see [[AssayDesign]]
  */
-export function getById(options: IGetByIdOptions): void {
-
-    let config: IGetAssaysOptions = options;
-
-    if (arguments.length > 1) {
-        config = {
-            success: arguments[0],
-            failure: arguments[1],
-            parameters: { id: arguments[2] },
-            containerPath: arguments[3]
-        };
-    }
-
-    moveParameter(config, 'id'); // repeat?
-    getAssays(config);
+export function getById(options: GetByIdOptions): XMLHttpRequest {
+    return getAssays(applyArguments(arguments, options, 'id'));
 }
 
-export interface IGetByNameOptions extends IGetAssaysOptions {
-    /**String name of the assay.*/
+export interface GetByNameOptions extends GetAssaysOptions {
+    /** String name of the assay. */
     name: string
 }
 
@@ -195,25 +191,12 @@ export interface IGetByNameOptions extends IGetAssaysOptions {
  * @param options
  * @see [[AssayDesign]]
  */
-export function getByName(options: IGetByNameOptions): void {
-
-    let config: IGetAssaysOptions = options;
-
-    if (arguments.length > 1) {
-        config = {
-            success: arguments[0],
-            failure: arguments[1],
-            parameters: { name: arguments[2] },
-            containerPath: arguments[3]
-        };
-    }
-
-    moveParameter(config, 'name'); // repeat?
-    getAssays(config);
+export function getByName(options: GetByNameOptions): XMLHttpRequest {
+    return getAssays(applyArguments(arguments, options, 'name'));
 }
 
-export interface IGetByTypeOptions extends IGetAssaysOptions {
-    /**String name of the assay type.  "ELISpot", for example.*/
+export interface GetByTypeOptions extends GetAssaysOptions {
+    /** String name of the assay type (e.g. "ELISpot"). */
     type: string
 }
 
@@ -222,59 +205,40 @@ export interface IGetByTypeOptions extends IGetAssaysOptions {
  * @param options
  * @see [[AssayDesign]]
  */
-export function getByType(options: IGetByTypeOptions): void {
-
-    let config: IGetAssaysOptions = options;
-
-    if (arguments.length > 1) {
-        config = {
-            success: arguments[0],
-            failure: arguments[1],
-            parameters: { type: arguments[2] },
-            containerPath: arguments[3]
-        };
-    }
-
-    moveParameter(config, 'type'); // repeat?
-    getAssays(config);
+export function getByType(options: GetByTypeOptions): XMLHttpRequest {
+    return getAssays(applyArguments(arguments, options, 'type'));
 }
 
-export interface IGetNAbRunsOptions {
-    /**The name of the NAb assay design for which runs are to be retrieved.*/
+export interface GetNAbRunsOptions extends RequestCallbackOptions {
+    /** The name of the NAb assay design for which runs are to be retrieved. */
     assayName: string
-    /**Whether neutralization should be calculated on the server.*/
+    /** Whether neutralization should be calculated on the server. */
     calculateNeut?: boolean
-    /**The path to the container in which the schema and query are defined,
-     * if different than the current container. If not supplied, the current container's path will be used.*/
-    containerPath?: string
     /**
-     * Function called when execution of the "getNAbRuns" function fails.
-     * This function will be called with the following arguments:
-     * * errorInfo: an object describing the error with the following fields:
-     *   * exception: the exception message
-     *   * exceptionClass: the Java class of the exception thrown on the server
-     *   * stackTrace: the Java stack trace at the point when the exception occurred
-     * * responseObj: the XMLHttpResponseObject instance used to make the AJAX request
-     * * options: the options used for the AJAX request
+     * The path to the container in which the schema and query are defined, if different than the current container.
+     * If not supplied, the current container's path will be used.
      */
-    failure?: (errorInfo: {exception: string, exceptionClass: any, stackTrace: any}, responseObj: object, options: object) => any
-    /**Array of objects created by {@link LABKEY.Filter.create}.*/
-    filterArray?: Array<IFilter>
+    containerPath?: string
+    /** Array of objects created by {@link LABKEY.Filter.create}. */
+    filterArray?: IFilter[]
     /**Whether the parameters used in the neutralization curve fitting calculation
      * should be included in the response.*/
     includeFitParameters?: boolean
     /**Whether or not statistics (standard deviation, max, min, etc.) should
      * be returned with calculations and well data.*/
     includeStats?: boolean
-    /**Whether well-level data should be included in the response.*/
+    /** Whether well-level data should be included in the response. */
     includeWells?: boolean
-    /**The maximum number of runs to return from the server (defaults to 100).
-     * If you want to return all possible rows, set this config property to -1.*/
+    /**
+     * The maximum number of runs to return from the server (defaults to 100).
+     * If you want to return all possible rows, set this config property to -1.
+     */
     maxRows?: number
-    /**The index of the first row to return from the server (defaults to 0).
-     * Use this along with the maxRows config property to request pages of data.*/
+    /**
+     * The index of the first row to return from the server (defaults to 0).
+     * Use this along with the maxRows config property to request pages of data.
+     */
     offset?: number
-    scope?: any
     /**
      * String description of the sort.  It includes the column names
      * listed in the URL of a sorted data region (with an optional minus prefix to indicate
@@ -283,15 +247,9 @@ export interface IGetNAbRunsOptions {
      */
     sort?: string
     /**
-     * Function called when the "getNAbRuns" function executes successfully.
-     * This function will be called with the following arguments:
-     * * runs: an array of NAb run objects
-     * * options: the options used for the AJAX request
-     * * responseObj: the XMLHttpResponseObject instance used to make the AJAX request
+     * The maximum number of milliseconds to allow for this operation before generating
+     * a timeout error (defaults to 30000).
      */
-    success: Function
-    /**The maximum number of milliseconds to allow for this operation before
-     * generating a timeout error (defaults to 30000).*/
     timeout?: number
 }
 
@@ -299,8 +257,7 @@ export interface IGetNAbRunsOptions {
  * Select NAb assay data from an assay folder.
  * @param options
  */
-export function getNAbRuns(options: IGetNAbRunsOptions): void {
-
+export function getNAbRuns(options: GetNAbRunsOptions): XMLHttpRequest {
     let params = merge({}, options);
 
     if (options.sort) {
@@ -318,68 +275,57 @@ export function getNAbRuns(options: IGetNAbRunsOptions): void {
         }
     }
 
-    const success = getOnSuccess(options);
-
-    request({
-        url: buildURL('nabassay', 'getNabRuns', options.containerPath),
-        method: 'GET',
+    return request({
+        url: buildURL('nabassay', 'getNabRuns.api', options.containerPath),
         params: appendFilterParams(params, options.filterArray),
-        success: getCallbackWrapper(function(data: any) {
-            if (success) {
-                success.call(options.scope, data.runs);
-            }
-        }, this),
+        success: getCallbackWrapper(
+            getOnSuccess(options),
+            options.scope,
+            false,
+            (data) => data.runs
+        ),
         failure: getCallbackWrapper(getOnFailure(options), options.scope, true),
         timeout: options.timeout
     });
 }
 
-export interface IGetStudyNabGraphURLOptions {
-    /**The desired title for the chart. Defaults to no title.*/
+export interface GetStudyNabGraphURLResponse {
+    /** IDs of the samples that were successfully graphed. */
+    objectIds: any[]
+    /** URL of the dilution curve graph. */
+    url: string
+}
+
+export interface GetStudyNabGraphURLOptions extends RequestCallbackOptions<GetStudyNabGraphURLResponse> {
+    /** The desired title for the chart. Defaults to no title. */
     chartTitle?: string
-    /**The path to the study container containing the NAb summary data,
-     * if different than the current container. If not supplied, the current container's path will be used.*/
+    /**
+     * The path to the study container containing the NAb summary data, if different than the current container.
+     * If not supplied, the current container's path will be used.
+     */
     containerPath?: string
-    /**
-     * Function called when execution of the "getStudyNabGraphURL" function fails.
-     * This function will be called with the following arguments:
-     * * errorInfo: an object describing the error with the following fields:
-     *   * exception: the exception message
-     *   * exceptionClass: the Java class of the exception thrown on the server
-     *   * stackTrace: the Java stack trace at the point when the exception occurred
-     * * responseObj: the XMLHttpResponseObject instance used to make the AJAX request
-     * * options: the options used for the AJAX request
-     */
-    failure?: Function
-    /**Allowable values are FIVE_PARAMETER, FOUR_PARAMETER, and POLYNOMIAL.
-     * Defaults to FIVE_PARAMETER.*/
+    /** Allowable values are FIVE_PARAMETER, FOUR_PARAMETER, and POLYNOMIAL. Defaults to FIVE_PARAMETER. */
     fitType?: 'FIVE_PARAMETER' | 'FOUR_PARAMETER' | 'POLYNOMIAL'
-    /**Desired height of the graph image in pixels. Defaults to 300.*/
+    /** Desired height of the graph image in pixels. Defaults to 300. */
     height?: number
-    /**The object Ids for the NAb data rows that have been copied to the study.
-     * This method will ignore requests to graph any object IDs that the current user does not have permission to view.*/
-    objectIds: Array<string | number>
-    scope?: any
     /**
-     * Function called when the "getStudyNabGraphURL" function executes successfully.
-     * This function will be called with the following arguments:
-     * * result: an object with the following properties:
-     *   * url: a string URL of the dilution curve graph.
-     *   * objectIds: an array containing the IDs of the samples that were successfully graphed.
-     * * responseObj: the XMLHttpResponseObject instance used to make the AJAX request
-     * * options: the options used for the AJAX request
+     * The object Ids for the NAb data rows that have been copied to the study.
+     * This method will ignore requests to graph any object IDs that the current
+     * user does not have permission to view.
      */
-    success: Function
-    /**The maximum number of milliseconds to allow for this operation before
-     * generating a timeout error (defaults to 30000).*/
+    objectIds: Array<string | number>
+    /**
+     * The maximum number of milliseconds to allow for this operation before generating
+     * a timeout error (defaults to 30000).
+     */
     timeout?: number
-    /**Desired width of the graph image in pixels. Defaults to 425.*/
+    /** Desired width of the graph image in pixels. Defaults to 425. */
     width?: number
 }
 
 /**
- * Retrieve the URL of an image that contains a graph of dilution curves for NAb results that have been copied to a study.
- * Note that this method must be executed against the study folder containing the copied NAb summary data.
+ * Retrieve the URL of an image that contains a graph of dilution curves for NAb results that have been copied
+ * to a study. Note that this method must be executed against the study folder containing the copied NAb summary data.
  * #### Examples
  *
  * ```
@@ -408,16 +354,14 @@ export interface IGetStudyNabGraphURLOptions {
  * </script>
  * <div id="graphDiv">
  * ```
- * @param {IGetStudyNabGraphURLOptions} options
+ * @param options
  */
-export function getStudyNabGraphURL(options: IGetStudyNabGraphURLOptions): void {
-
+export function getStudyNabGraphURL(options: GetStudyNabGraphURLOptions): XMLHttpRequest {
     let params: any = {};
     applyTranslated(params, options, { objectIds: 'id' }, true, false);
 
-    request({
-        url: buildURL('nabassay', 'getStudyNabGraphURL'),
-        method: 'GET',
+    return request({
+        url: buildURL('nabassay', 'getStudyNabGraphURL.api', options.containerPath),
         params,
         success: getCallbackWrapper(getOnSuccess(options), options.scope),
         failure: getCallbackWrapper(getOnFailure(options) || displayAjaxErrorResponse, options.scope, true),
@@ -425,41 +369,32 @@ export function getStudyNabGraphURL(options: IGetStudyNabGraphURLOptions): void 
     });
 }
 
-export interface IGetStudyNabRunsOptions {
-    /**Whether neutralization should be calculated on the server.*/
+export interface GetStudyNabRunsOptions extends RequestCallbackOptions {
+    /** Whether neutralization should be calculated on the server. */
     calculateNeut?: boolean
-    /**The path to the study container containing the NAb summary,
-     * if different than the current container. If not supplied, the current container's path will be used.*/
-    containerPath?: string
-    /**Function called when execution of the "getStudyNabRuns" function fails.
-     * This function will be called with the following arguments:
-     * * errorInfo: an object describing the error with the following fields:
-     *   * exception: the exception message
-     *   * exceptionClass: the Java class of the exception thrown on the server
-     *   * stackTrace: the Java stack trace at the point when the exception occurred
-     * * responseObj: the XMLHttpResponseObject instance used to make the AJAX request
-     * * options: the options used for the AJAX request
-     */
-    failure?: Function
-    /**Whether the parameters used in the neutralization curve fitting calculation
-     * should be included in the response.*/
-    includeFitParameters?: boolean
-    /**Whether or not statistics (standard deviation, max, min, etc.) should
-     * be returned with calculations and well data.*/
-    includeStats?: boolean
-    /**Whether well-level data should be included in the response.*/
-    includeWells?: boolean
-    /**The object Ids for the NAb data rows that have been copied to the study.*/
-    objectIds: Array<string | number>
-    scope?: any
     /**
-     * Function called when the "getStudyNabRuns" function executes successfully.
-     * This function will be called with the following arguments:
-     * * runs: an array of NAb run objects
+     * The path to the study container containing the NAb summary, if different than the current container.
+     * If not supplied, the current container's path will be used.
      */
-    success: Function
-    /**The maximum number of milliseconds to allow for this operation before
-     * generating a timeout error (defaults to 30000).*/
+    containerPath?: string
+    /**
+     * Whether the parameters used in the neutralization curve fitting calculation should be included
+     * in the response.
+     */
+    includeFitParameters?: boolean
+    /**
+     * Whether or not statistics (standard deviation, max, min, etc.) should be returned with calculations
+     * and well data.
+     */
+    includeStats?: boolean
+    /** Whether well-level data should be included in the response. */
+    includeWells?: boolean
+    /** The object Ids for the NAb data rows that have been copied to the study. */
+    objectIds: Array<string | number>
+    /**
+     * The maximum number of milliseconds to allow for this operation before generating
+     * a timeout error (defaults to 30000).
+     */
     timeout?: number
 }
 
@@ -468,20 +403,16 @@ export interface IGetStudyNabRunsOptions {
  * method must be executed against the study folder containing the copied NAb summary data.
  * @param options
  */
-export function getStudyNabRuns(options: IGetStudyNabRunsOptions): void {
-
-    let params = merge({}, options);
-    const success = getOnSuccess(options);
-
-    request({
-        url: buildURL('nabassay', 'getStudyNabRuns', options.containerPath),
-        method: 'GET',
-        params,
-        success: getCallbackWrapper(function(data: any) {
-            if (success) {
-                success.call(options.scope, data.runs);
-            }
-        }, this),
+export function getStudyNabRuns(options: GetStudyNabRunsOptions): XMLHttpRequest {
+    return request({
+        url: buildURL('nabassay', 'getStudyNabRuns.api', options.containerPath),
+        params: merge({}, options),
+        success: getCallbackWrapper(
+            getOnSuccess(options),
+            options.scope,
+            false,
+            (data) => data.runs
+        ),
         failure: getCallbackWrapper(getOnFailure(options) || displayAjaxErrorResponse, options.scope, true),
         timeout: options.timeout
     });
@@ -491,20 +422,7 @@ export function getStudyNabRuns(options: IGetStudyNabRunsOptions): void {
  * @hidden
  * @private
  */
-function getSuccessCallbackWrapper(success: Function, scope: any): AjaxHandler {
-    return getCallbackWrapper(function(data: any, response: any) {
-        if (success) {
-            success.call(this, data.definitions, response);
-        }
-    }, (scope || this));
-}
-
-/**
- * @hidden
- * @private
- */
 function moveParameter(config: any, param: any): void {
-
     if (!config.parameters) {
         config.parameters = {};
     }